Polybench: rename jacobi-2d-imper -> jacobi-2d
