+++ SingleSource/Benchmarks/Polybench/stencils/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(adi)
 add_subdirectory(fdtd-2d)
 add_subdirectory(jacobi-1d)
-add_subdirectory(jacobi-2d-imper)
+add_subdirectory(jacobi-2d)
 add_subdirectory(seidel-2d)
