+++ SingleSource/Benchmarks/Polybench/stencils/jacobi-2d/jacobi-2d.reference_output.small
